// contracts/typescript/models/InvitationPage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HomeOfficeReservation API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { InvitationProfile } from './InvitationProfile';
import {
    InvitationProfileFromJSON,
    InvitationProfileFromJSONTyped,
    InvitationProfileToJSON,
    InvitationProfileToJSONTyped,
} from './InvitationProfile';

/**
 *
 * @export
 * @interface InvitationPage
 */
export interface InvitationPage {
    /**
     *
     */
    members: Array<InvitationProfile>;
    /**
     *
     */
    nextAfter: string | null;
}

/**
 * Check if a given object implements the InvitationPage interface.
 */
export function instanceOfInvitationPage(value: object): value is InvitationPage {
    if (!('members' in value) || value['members'] === undefined) return false;
    if (!('nextAfter' in value) || value['nextAfter'] === undefined) return false;
    return true;
}

export function InvitationPageFromJSON(json: any): InvitationPage {
    return InvitationPageFromJSONTyped(json, false);
}

export function InvitationPageFromJSONTyped(json: any, ignoreDiscriminator: boolean): InvitationPage {
    if (json == null) {
        return json;
    }
    return {

        'members': ((json['members'] as Array<any>).map(InvitationProfileFromJSON)),
        'nextAfter': json['nextAfter'],
    };
}

export function InvitationPageToJSON(json: any): InvitationPage {
    return InvitationPageToJSONTyped(json, false);
}

export function InvitationPageToJSONTyped(value?: InvitationPage | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'members': ((value['members'] as Array<any>).map(InvitationProfileToJSON)),
        'nextAfter': value['nextAfter'],
    };
}